Changed warning level to 4 on MSVC, with minor refactors

+++ include/memory/memory_common.h
@@ -3,7 +3,7 @@
 namespace kab
 {
 	using size_t = decltype(sizeof(void*));
-	inline constexpr size_t size_t_max_v = -1;
+	inline constexpr size_t size_t_max_v = static_cast<size_t>(-1);
 
 	// assume 64 bits for now
 #	define KAB_ATOMIC_FETCH_INC_SIZE_T_RELAXED(v) KAB_ATOMIC_FETCH_INC_UINT64_RELAXED(v)
